test(editor): split authored script diagnostics fixture
